@@ -1,11 +1,10 @@
-﻿namespace InfluxGateway
+﻿namespace InfluxGateway;
+
+public interface IInfluxConnectionSettings
 {
-    public interface IInfluxConnectionSettings
-    {
-        string InfluxUrl { get; }
-        string InfluxUsername { get; }
-        string InfluxPassword { get; }
-        string InfluxDatabaseConnection { get; }
-        string InfluxQuery { get; }
-    }
+    string InfluxUrl { get; }
+    string InfluxUsername { get; }
+    string InfluxPassword { get; }
+    string InfluxDatabaseConnection { get; }
+    string InfluxQuery { get; }
 }
