@@ -1,54 +1,45 @@
-﻿using System.Linq;
-using System.Threading.Tasks;
-using InfluxDB.Net;
-using InfluxGateway.Controllers;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
+﻿
+namespace InfluxGateway;
 
-namespace InfluxGateway
+public class InfluxDatabase : IInfluxDatabase
 {
-    public class InfluxDatabase : IInfluxDatabase
+    private readonly InfluxDb _influxDb;
+    private readonly ILogger<InfluxController> _logger;
+    private readonly IConfiguration _configuration;
+    private readonly IInfluxConnectionSettings _influxConnectionSettings;
+
+    public InfluxDatabase(IConfiguration configuration, ILogger<InfluxController> logger, IInfluxConnectionSettings influxConnectionSettings)
     {
-
-        private InfluxDb InfluxDb { get; }
-        private IConfiguration Configuration { get; }
-        private ILogger<InfluxController> Logger { get; }
-        private IInfluxConnectionSettings InfluxConnectionSettings { get; }
-
-        public InfluxDatabase(IConfiguration configuration, ILogger<InfluxController> logger, IInfluxConnectionSettings influxConnectionSettings)
-        {
-            Configuration = configuration;
-            Logger = logger;
-            InfluxConnectionSettings = influxConnectionSettings;
-            InfluxDb = GetInfluxConnection();
-
-        }
-
-        /// <summary>
-        /// TODO: Sanitise user input to prevent command injection.
-        /// </summary>
-        /// <param name="sensorName"></param>
-        /// <returns></returns>
-        private string BuildQuery(string sensorName)
-        {
-            return InfluxConnectionSettings.InfluxQuery.Replace("%s", sensorName);
-        }
-
-        public async Task<string> GetFirstResultForInfluxQuery(string query)
-        {
-            var x = await InfluxDb.QueryAsync(InfluxConnectionSettings.InfluxDatabaseConnection, BuildQuery(query));
-            return x.FirstOrDefault().Values[0][1].ToString();
-        }
-
-        /// <summary>
-        /// Construct a Influx Client
-        /// </summary>
-        /// <returns></returns>
-        private InfluxDb GetInfluxConnection()
-        {
-            Logger.LogDebug("New connection to be created against {url}", InfluxConnectionSettings.InfluxUrl);
-            return new InfluxDb(InfluxConnectionSettings.InfluxUrl, InfluxConnectionSettings.InfluxUsername, InfluxConnectionSettings.InfluxPassword);
-        }
-
+        _configuration = configuration;
+        _logger = logger;
+        _influxConnectionSettings = influxConnectionSettings;
+        _influxDb = GetInfluxConnection();
     }
+
+    /// <summary>
+    /// TODO: Sanitise user input to prevent command injection.
+    /// </summary>
+    /// <param name="sensorName"></param>
+    /// <returns></returns>
+    private string BuildQuery(string sensorName)
+    {
+        return _influxConnectionSettings.InfluxQuery.Replace("%s", sensorName);
+    }
+
+    public async Task<string> GetFirstResultForInfluxQuery(string query)
+    {
+        var x = await _influxDb.QueryAsync(_influxConnectionSettings.InfluxDatabaseConnection, BuildQuery(query));
+        return x.FirstOrDefault().Values[0][1].ToString();
+    }
+
+    /// <summary>
+    /// Construct a Influx Client
+    /// </summary>
+    /// <returns></returns>
+    private InfluxDb GetInfluxConnection()
+    {
+        _logger.LogDebug("New connection to be created against {url}", _influxConnectionSettings.InfluxUrl);
+        return new InfluxDb(_influxConnectionSettings.InfluxUrl, _influxConnectionSettings.InfluxUsername, _influxConnectionSettings.InfluxPassword);
+    }
+
 }
